Modify libike to cast a value used in an OpenSSL function to avoid a compiler warning on x64 Windows builds.
git-svn-id: svn://svn.shrew.net/ike/head@774 1bb156eb-19af-db11-b919-0014c23ffdf2

+++ source/libike/config.file.cpp
@@ -483,7 +483,7 @@ bool _CONFIG_MANAGER::file_load_pcf( CONFIG * config, const char * path, bool & 
 				&ctx_cipher,
 				pwd.buff(),
 				pwd.buff(),
-				pwd.size() );
+				( unsigned int ) pwd.size() );
 
 			pwlen -= pwd.buff()[ pwd.size() - 1 ];
 			pwd.size( pwlen );
